# docs/openapi/autoflow-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoFlow API",
    "version": "v1",
    "description": "Contrato OpenAPI versionado da API AutoFlow. Exportado a partir do contrato SpringDoc da aplicação e complementado com parâmetros de execução, autenticação e exemplos em docs/openapi/README.md."
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "API local (backend)"
    },
    {
      "url": "http://localhost:4200/api",
      "description": "API local pelo proxy do frontend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/clientes/me": {
      "get": {
        "operationId": "meuPerfilCliente",
        "summary": "Consultar meu perfil",
        "tags": [
          "clientes"
        ],
        "description": "Consultar meu perfil Permissão: CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/clientes": {
      "post": {
        "operationId": "cadastrarCliente",
        "summary": "Cadastrar cliente",
        "tags": [
          "clientes"
        ],
        "description": "Cadastrar cliente",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listarClientes",
        "summary": "Listar clientes",
        "tags": [
          "clientes"
        ],
        "description": "Listar clientes Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "operationId": "consultarCliente",
        "summary": "Consultar cliente",
        "tags": [
          "clientes"
        ],
        "description": "Consultar cliente Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 12345678901
          }
        ]
      },
      "delete": {
        "operationId": "deletarCliente",
        "summary": "Excluir cliente",
        "tags": [
          "clientes"
        ],
        "description": "Excluir cliente Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "cliente deletado com sucesso"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      }
    },
    "/clientes/{id}/atualizacao": {
      "patch": {
        "operationId": "atualizarCliente",
        "summary": "Atualizar cliente",
        "tags": [
          "clientes"
        ],
        "description": "Atualizar cliente Permissão: ADMIN ou ATENDENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteRequest"
              }
            }
          }
        }
      }
    },
    "/clientes/{id}/status": {
      "patch": {
        "operationId": "alterarStatusCliente",
        "summary": "Alterar status do cliente",
        "tags": [
          "clientes"
        ],
        "description": "Ativar ou desativar cliente. Permissão: ADMIN.",
        "responses": {
          "200": {
            "description": "Status do cliente alterado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteStatusRequest"
              },
              "examples": {
                "inativar": {
                  "value": {
                    "status": "INATIVO"
                  }
                },
                "ativar": {
                  "value": {
                    "status": "ATIVO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token retornado por POST /auth/login no header Authorization: Bearer <token>."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida, JSON inválido, validação ou regra de negócio.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "JSON inválido"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "JWT ausente/inválido ou token público ausente, inválido ou expirado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Token inválido ou expirado"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Usuário autenticado sem a role necessária.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Acesso negado"
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso não encontrado"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflito de unicidade ou recurso já cadastrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso já cadastrado"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string",
            "example": "Recurso não encontrado"
          },
          "campos": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "ClienteRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "cpfCnpj": {
            "type": "string",
            "minLength": 11,
            "maxLength": 14,
            "example": "12345678901"
          },
          "telefone": {
            "type": "string",
            "example": "11999999999"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@example.com"
          }
        },
        "required": [
          "nome",
          "cpfCnpj",
          "telefone",
          "email"
        ]
      },
      "ClienteStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO"
            ],
            "example": "INATIVO"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "VeiculoClienteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "marca": {
            "type": "string",
            "example": "Toyota"
          },
          "ano": {
            "type": "integer",
            "format": "int64",
            "example": 2020
          },
          "placa": {
            "type": "string",
            "example": "ABC1D23"
          },
          "modelo": {
            "type": "string",
            "example": "Corolla"
          }
        }
      },
      "ClienteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Maria da Silva"
          },
          "cpfCnpj": {
            "type": "string",
            "example": "12345678901"
          },
          "telefone": {
            "type": "string",
            "example": "11999999999"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@example.com"
          },
          "status": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO"
            ],
            "example": "ATIVO"
          },
          "veiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VeiculoClienteResponse"
            }
          }
        }
      },
...
    }
  }
}